requirements.txt
flask==3.0.0
gunicorn==21.2.0
opencv-python-headless==4.8.1.78
ultralytics==8.1.0
numpy==1.24.3
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0